launcher: split backend_mgr into a package with its settings-change half
backend_mgr.py had grown to hold two unrelated jobs: process lifecycle/UI, and
working out what a settings change means for the backend. Split along that seam,
using the same package shape as hud_mgr:

  backend_mgr/__init__.py        re-exports BackendAppMgr
  backend_mgr/backend_mgr.py     lifecycle + buttons + manual save
  backend_mgr/settings_change.py BackendSettingsChangeBase

settings_change.py is a real intermediate subclass of PngAppMgrBase rather than a
standalone mixin, so self.debug_log / self.curr_settings resolve through the MRO to
their one real definition - go-to-definition and find-references land in base_mgr.py
instead of on re-declared stubs.

That needed an opt-out from the identity check in __init_subclass__, since an
intermediate class has no MODULE_PATH/DISPLAY_NAME/SHORT_NAME of its own; the
concrete subclass still has to supply them. It is a class attribute read off
cls.__dict__ (so it cannot inherit) rather than the more natural class keyword
argument, because the Shiboken metaclass behind QObject rejects those.

No behaviour change. Nothing outside the package moved: subsystems/__init__.py
still imports BackendAppMgr from .backend_mgr, and the PyInstaller spec picks the
new module up via collect_submodules.
